pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "devspec"
version = "0.1.0"
description = "Spec-driven development workflow engine for Claude Code"
readme = "README.md"
license = "MIT"
requires-python = ">=3.14"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.scripts]
devspec = "devspec.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
devspec = ["data/**/*"]

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
